+++ tests/conftest.py
@@ -36,7 +36,7 @@ from dodal.devices.backlight import Backlight
 from dodal.devices.detector.detector_motion import DetectorMotion
 from dodal.devices.eiger import EigerDetector
 from dodal.devices.fast_grid_scan import FastGridScanCommon
-from dodal.devices.i03 import Beamstop
+from dodal.devices.i03 import Beamstop, BeamstopPositions
 from dodal.devices.i03.dcm import DCM
 from dodal.devices.oav.oav_detector import OAVConfig
 from dodal.devices.oav.pin_image_recognition import PinTipDetection
@@ -1501,3 +1501,24 @@ def xbpm_and_transmission_wrapper_composite(
     return XBPMAndTransmissionWrapperComposite(
         undulator, xbpm_feedback, attenuator, dcm
     )
+
+
+@pytest.fixture
+def beamstop_i03(
+    beamline_parameters: GDABeamlineParameters, sim_run_engine: RunEngineSimulator
+) -> Generator[Beamstop, Any, Any]:
+    with patch(
+        "dodal.beamlines.i03.get_beamline_parameters", return_value=beamline_parameters
+    ):
+        beamstop = i03.beamstop(connect_immediately=True, mock=True)
+        patch_motor(beamstop.x_mm)
+        patch_motor(beamstop.y_mm)
+        patch_motor(beamstop.z_mm)
+        set_mock_value(beamstop.x_mm.user_readback, 1.52)
+        set_mock_value(beamstop.y_mm.user_readback, 44.78)
+        set_mock_value(beamstop.z_mm.user_readback, 30.0)
+        sim_run_engine.add_read_handler_for(
+            beamstop.selected_pos, BeamstopPositions.DATA_COLLECTION
+        )
+        yield beamstop
+        beamline_utils.clear_devices()
