+++ tests/unit_tests/hyperion/conftest.py
@@ -1,18 +1,11 @@
-from collections.abc import Generator
 from functools import partial
 from importlib import resources
 from pathlib import Path
-from typing import Any
 from unittest.mock import MagicMock, patch
 
 import pytest
 from bluesky.run_engine import RunEngine
-from bluesky.simulators import RunEngineSimulator
 from dodal.beamlines import i03
-from dodal.common.beamlines import beamline_utils
-from dodal.common.beamlines.beamline_parameters import (
-    GDABeamlineParameters,
-)
 from dodal.devices.aperturescatterguard import (
     ApertureScatterguard,
 )
@@ -21,7 +14,7 @@ from dodal.devices.backlight import Backlight
 from dodal.devices.detector.detector_motion import DetectorMotion
 from dodal.devices.eiger import EigerDetector
 from dodal.devices.flux import Flux
-from dodal.devices.i03 import Beamstop, BeamstopPositions
+from dodal.devices.i03 import Beamstop
 from dodal.devices.i03.dcm import DCM
 from dodal.devices.oav.oav_detector import OAV
 from dodal.devices.oav.oav_parameters import OAVParameters
@@ -30,7 +23,6 @@ from dodal.devices.s4_slit_gaps import S4SlitGaps
 from dodal.devices.smargon import Smargon
 from dodal.devices.synchrotron import Synchrotron
 from dodal.devices.undulator import Undulator
-from dodal.devices.util.test_utils import patch_motor
 from dodal.devices.xbpm_feedback import XBPMFeedback
 from dodal.devices.zebra.zebra import Zebra
 from dodal.devices.zebra.zebra_controlled_shutter import ZebraShutter
@@ -183,27 +175,6 @@ def test_multi_rotation_params():
             "tests/test_data/parameter_json_files/good_test_multi_rotation_scan_parameters.json"
         )
     )
-
-
-@pytest.fixture
-def beamstop_i03(
-    beamline_parameters: GDABeamlineParameters, sim_run_engine: RunEngineSimulator
-) -> Generator[Beamstop, Any, Any]:
-    with patch(
-        "dodal.beamlines.i03.get_beamline_parameters", return_value=beamline_parameters
-    ):
-        beamstop = i03.beamstop(connect_immediately=True, mock=True)
-        patch_motor(beamstop.x_mm)
-        patch_motor(beamstop.y_mm)
-        patch_motor(beamstop.z_mm)
-        set_mock_value(beamstop.x_mm.user_readback, 1.52)
-        set_mock_value(beamstop.y_mm.user_readback, 44.78)
-        set_mock_value(beamstop.z_mm.user_readback, 30.0)
-        sim_run_engine.add_read_handler_for(
-            beamstop.selected_pos, BeamstopPositions.DATA_COLLECTION
-        )
-        yield beamstop
-        beamline_utils.clear_devices()
 
 
 @pytest.fixture
