+++ tests/unit_tests/hyperion/experiment_plans/test_hyperion_flyscan_xray_centre_plan.py
@@ -109,7 +109,7 @@ class TestFlyscanXrayCentrePlan:
         autospec=True,
     )
     @patch(
-        "mx_bluesky.hyperion.experiment_plans.change_aperture_then_move_plan.move_x_y_z",
+        "mx_bluesky.common.experiment_plans.change_aperture_then_move_plan.move_x_y_z",
         autospec=True,
     )
     @pytest.mark.skip(
@@ -166,7 +166,7 @@ class TestFlyscanXrayCentrePlan:
         autospec=True,
     )
     @patch(
-        "mx_bluesky.hyperion.experiment_plans.change_aperture_then_move_plan.move_x_y_z",
+        "mx_bluesky.common.experiment_plans.change_aperture_then_move_plan.move_x_y_z",
         autospec=True,
     )
     async def test_when_gridscan_finished_then_dev_shm_disabled(
